chore: add SSH auto login script using 1Password
This commit adds a script called `ssh_auto_login.exp` that can be used to automate SSH login using 1Password for password management. The script uses the 'expect' command to automate interactions with the SSH login process.

The script accepts two arguments:
1. The 1Password key reference for the SSH password (e.g., "op://Work Account/SSH RUMBA UTIL/password")
2. The SSH host name as defined in the SSH config file (e.g., "rumbautil")

The script retrieves the password from 1Password using the `op read` command and assigns it to the `password` variable. It then starts the SSH process for the specified host using the `spawn` command. When prompted for a password, the script provides the retrieved password using the `send` command.

To run this script, use the following command:
```
./ssh_auto_login.exp "op://Work Account/